Add a role var to determine if systemd automation foo should be installed
The role variable defaults to false to maintain compatibility with the
role as it currently exists.

Also create a new systemd_automation scenario where the role variable
is set to true and move the systemd-specific Molecule tests there.
