+++ molecule/systemd_automation/INSTALL.rst
@@ -1,0 +1,1 @@
+../default/INSTALL.rst
