+++ molecule/systemd_automation/converge.yml
@@ -1,0 +1,18 @@
+---
+- name: Converge
+  hosts: all
+  tasks:
+    - name: Include ansible-role-cyhy-reports
+      # We do prepend the name of the role to the role variables, but
+      # Molecule does its own role discovery with inconsistent naming.
+      # This is the reason for the noqa below.
+      ansible.builtin.include_role: # noqa var-naming[no-role-prefix]
+        name: ansible-role-cyhy-reports
+      vars:
+        cyhy_reports_automation: true
+        cyhy_reports_maxmind_account_id: >-
+          {{ lookup('aws_ssm', '/cyhy/core/geoip/account_id',
+          region='us-east-1') }}
+        cyhy_reports_maxmind_license_key: >-
+          {{ lookup('aws_ssm', '/cyhy/core/geoip/license_key',
+          region='us-east-1') }}
