+++ molecule/systemd_automation/externally-managed-python.yml
@@ -1,0 +1,1 @@
+../default/externally-managed-python.yml
