+++ molecule/systemd_automation/prepare.yml
@@ -1,0 +1,1 @@
+../default/prepare.yml
