+++ molecule/systemd_automation/requirements.yml
@@ -1,0 +1,1 @@
+../default/requirements.yml
