+++ molecule/systemd_automation/templates/EXTERNALLY-MANAGED.j2
@@ -1,0 +1,1 @@
+../../default/templates/EXTERNALLY-MANAGED.j2
