+++ molecule/systemd_automation/tests/conftest.py
@@ -1,0 +1,1 @@
+../../default/tests/conftest.py
