+++ molecule/systemd_automation/tests/test_default.py
@@ -1,0 +1,1 @@
+../../default/tests/test_default.py
