+++ molecule/systemd_automation/upgrade.yml
@@ -1,0 +1,1 @@
+../default/upgrade.yml
