+++ tasks/main.yml
@@ -119,3 +119,5 @@
 - name: Install systemd timer and service to generate CyHy reports
   ansible.builtin.include_tasks:
     file: setup_automated_cyhy_reporting.yml
+  when:
+    - cyhy_reports_automation
